requirements.txt
transformers
pydub
faster_whisper
soundfile
faster_whisper
pickle
gradio
pyautogui
pillow 
timm